web/src/requirements.txt
pyramid==1.10
pyramid-jinja2
mysql-connector-python
python-dotenv
bjoern
